requirements.txt
numpy<2
torch
torchvision
fastapi
uvicorn[standard]
pillow
groundingdino-py
sam2
huggingface-hub
